refactor(sidebar): extract thr user data to be a hook to use it in a lot of places
