+++ src/utils/store.jsx
@@ -1,6 +1,8 @@
+import { doc, getDoc } from "firebase/firestore"
 import { createContext, useContext, useEffect, useState } from "react"
+import { toast } from "react-toastify"
 
-import { auth } from "./firebase/config"
+import { auth, db } from "./firebase/config"
 
 export const StoreContext = createContext()
 
@@ -37,4 +39,28 @@ export const useAuth = () => {
     loading,
     isLoggedIn,
   }
+}
+
+export const useProfileData = () => {
+  const { user } = useAuth()
+  const [profileData, setProfileData] = useState(null)
+
+  useEffect(() => {
+    const fetchProfileData = async () => {
+      if (user) {
+        const docRef = doc(db, "users", user.uid)
+        const docSnap = await getDoc(docRef)
+
+        if (docSnap.exists()) {
+          setProfileData(docSnap.data())
+        } else {
+          toast.error("No such document!")
+        }
+      }
+    }
+
+    fetchProfileData()
+  }, [user])
+
+  return profileData
 }
